Increase ping timeout from 60 to 120 seconds
The Cirros image only sends out three DHCPDISCOVERs with 60 second
waits between. The ovs_neutron_agent may not have the port wired in
time for the first one, and with ping_timeout=60 the test will be
terminated before the second one is sent. This patch increases the
timeout to 120 seconds to prevent that.

Change-Id: I03fcec50b91fd89ad64e16fce82f9211c2e086a1
Partial-Bug: 1253896