Use service catalog_type when getting the heat/glance clients
The 'orchestration' service_type is currently hard-coded when getting
the heat client but there is an orchestration.catalog_type config option
available, so use that rather than hard-code the value.

The same is true for the _get_image_client method and
CONF.image.catalog_type option, so fix that as well.

Change-Id: I6b4ba44619b9ddf99f063bc28c7c4bbae67408a8