tighten up isolated creds create
we're spewing giant user and tenant names in logs, which make
actually getting to the bottom of things less useful, as we're
constantly doing left / right scrolling. Shorten this up.

this changes makes a number of fix ups here.

1) it shortens down the name to just classname + rand piece, no
need to add extra -tempest- -tenant, as it's pretty redundant. We
throw out any value which has a '.' in it, because that's not a
class name.

2) it fixes the scenario isolated creds call. __name__ is package
name, and cls.__name__ actually gives us the class name.

3) it stops using a variable in a different namespace as a
temporary variable for isolated creds create. This looks like it
was based on a find / replace refactor to remove importing modules.
It worked, but it was wrong.

4) it stops using the same random bits for user & tenant. There is
no reason these should match, and in the new scheme they would if
they shared random bits.

Change-Id: I216155d68cbdbf9bed10e2ab9b7a1ea714570358