Convert ironic, swift, and heat api tests to use global CONF object
This commit takes all the uses of config in the object_storage and
orchestration api tests and converts them to use the global CONF
object.

The _get_client_args() method from the base orchestration test class
is removed because it only uses the object attr config and isn't
actually used anywhere. So instead of converting it to use the global
object it is removed.

Partially implements bp config-cleanup

Change-Id: Ib8d7ad36047a1e752511bc4c5541d04ea61d6afc