Fix volume transfer tests with tenant isolation disabled
This commit fixes the volume transfer tests for when tenant isolation
is disabled. The volume transfer tests still had a reference to the old
auth style to get the tenant name if tenant isolation was disabled.
The recent auth_provider refactor didn't update this call so the test
would fail. This commit updates it to use the new credential dict to
get the tenant_name

Change-Id: Ie7fe39baf6f62fa6966631d7890891252602ed8d