Let the soft reboot really happen in the minimum scenario
The cirros image can support the soft reboot, but we reboot the machine
before the acpid started, it leads to additional 120 second wait before
nova falls back to the hard reboot.
As a solution we will ssh the machine before reboot, because the ssh
server starts after the acpid, so we will send the ACPI event, only
when the acpid is really running.

Change-Id: I3700a374dc937086518c5c843a288e935ab458bf