Fix test_create_list_delete_volume_transfer without tenant isolation
The test was making an assumption that there would be exactly one volume
transfer listed which is not the case when running without tenant
isolation. Note the test before it asserts the listed transfers are
greater than or equal to one so it works without tenant isolation.

The reason this doesn't work in the single-threaded case is probably
because another test is not cleaning up properly so there is a leak
somewhere, therefore this is a partial fix until the leak is resolved.

Partial-Bug: #1277788

Change-Id: I68b2df2d7792da7c3a9ce001af41cd199ccdbc13