Fix get versions call in verify_nova_api_versions()
This commit fixes the HTTP GET call to return the list of nova api
versions. The enpoint that the tempest clients use for servers
includes the api version. This is because nova registers each version
api as separate endpoints in keystone's catalog. So this commit fixes
the url to use the unversioned endpoint so the list of versions will
be returned.

Change-Id: Iaf4671d4d1dd285161aa5679648956c086a9119c