fix duplicate api_extensions
[network-feature-enabled] has 3 api_extensions entries in
tempest.conf.sample. 2 of them should belong to
[compute_features_group] and [volume_feature_group].

Generator compares the opt with opts in groups to find
the group in which the opt is. These 3 opts has same
'name', 'default value', and 'help message'. And first
match is '[network-feature-enabled]', so they go to the
same group.

This patch add some unique help info in api_extensions
opts to fix it.

Change-Id: I99dacc73281b16ed430912919defe3b4d9eb8b14
Closes-Bug: #1280085