XML endpoint client use correct string for enabled
The XML endpoint client was passing enabled='True' or
enabled='False' when creating or updating an endpoint.

The Keystone XML parser expects 'true' and 'false' for the
'enabled' attribute and only converts those strings to a Boolean
value, other values are passed through as strings.

Note that the XML identity_client correctly converts the Boolean
values to lowercase strings already.

Change-Id: I2f1374430d186959aca8fb90b4f7730044ec0c65
Closes-Bug: #1282266