Split out config option registration
This commit takes the option registration and local attr settings from
the tempest private config object and splits them out as separate
functions. This exposes a way to register the config options without
creating a config object. This enables the config options and defaults
to be more easily used for unit tests and external utilities, such as
the sample generator.

Change-Id: I2ed7e33a14e652f0ec9be2cce069b4925c02b3bd