Convert fake_config class to use config fixture
This commit changes the fake_config object used in the unit tests to
use the config fixture and use the real config defaults by registering
the config options and using the oslo config set_default method to
override the defaults when needed for testing. This will enable reusing
the tempest defaults in most places which should save effort duplicating
defaults in the fake object.

Partially Implements bp unit-tests

Change-Id: Ib9ee8e68f7000bfbcd3cd1c94cca103a3286ec77