Make reseller admin role configurable
This commit makes the reseller admin role configurable. Previously the
swift tests which used the reseller admin role had hardcoded the role
to be the default ResellerAdmin. However the role is configurable, so
this commit adds a new config option which can be used to set the role
to something besides the default.

Change-Id: I3296a343480d4a0e0e4099863eb8869ffb0f83c3