Unset the imageRef when booting from volume
When testing boot from volume, we were passing the image_ref no matter
what to servers.create. This will cause nova to use the metadata from
that image instead of the one that might contain the volume.

This patch sets the image to an empty string, which is the same
behaviour the nova CLI has when booting from a volume.

Change-Id: I59bd4e2be6d6fff782dd278390ad5241cd09e3df