ServerCfnInitTestJSON wait for stack complete before ssh
ssh connection has a default timeout of 300 seconds but
boot times of ~250 seconds have been observed so this is
failing for a proportion of runs.

This change waits for the entire stack to be in CREATE_COMPLETE
instead of just the server resource. This means the ssh daemon
will definitely be running when ssh is attempted.

Change-Id: Iab1c797b456c6bd46283a7348896a79ebc41d7e1
Closes-Bug: #1288993