Verify attributes through Nova list flavor-access API
This patch adds the JSON schema for Nova V2/V3 list flavor-access APIs
includes the attributes to block the backward incompatibility change
in the future.

The response body of v2 and v3 API is the below:
{
    "flavor_access": [
        {
            "flavor_id": "10"
            "tenant_id": "231a1e7fd1b344f9874d0334d98b459c"
        }
    ]
}

In addition, this patch doesn't change V3 test because it is skipped
owing to the bug.

Partially implements blueprint nova-api-attribute-test

Change-Id: I027dc96e5702cd03a7b85ba41c4c30215e321f30