Fix error trace induced by dhcp test actions
Once the agent is removed, we are unable to deallocate
dhcp resources for the port until the entire network is
torn down; this is fine, but by doing so we cause Neutron
Server to emit a log error.

This patch simply deletes the port before doing the
disassociation, however a cleaner way to deal with this
would be to avoid relying on port creation to trigger a
network scheduling, though, this would require a more
elaborated fix.

Closes-bug: 1292242

Change-Id: I187d8dd707d67eb39446a5558913331310bd1f47