Verify "create a server" API response attributes
Now most attributes of Nova v2/v3 APIs are not checked in Tempest,
and this patch adds some tests which check these attributes to block
the backward incompatibility change in the future.

This patch adds the checks of "create a server" API responses.

The response body of v2 API is the following:
  {
    "server": {
      "id": "d099f759-021f-41ad-8ad3-7d9ddecaf07a",
      "security_groups": [{"name": "default"}],
      "links": [
        {"href": "http://[..]", "rel": "self"},
        {"href": "http://[..]", "rel": "bookmark"}
      ],
      "adminPass": "6GLunfZpNbB8",
      "OS-DCF:diskConfig": "MANUAL"
    }
  }

The one of v3 API is the following:
  {
    "server": {
      "id": "c2f2ebe8-5a25-442b-83d2-9f0926a7a88a",
      "os-security-groups:security_groups": [{"name": "default"}],
      "links": [
        {"href": "http://[..]", "rel": "self"},
        {"href": "http://[..]", "rel": "bookmark"}
      ],
      "admin_password": "7XW3fhZpe6iF",
      "os-access-ips:access_ip_v6": "",
      "os-access-ips:access_ip_v4": "",
    }
  }

Partially implements blueprint nova-api-attribute-test

Change-Id: I584c47df3de8be2dd593b39b5486c9ed38c9a7ed