ssh instance validation add options for Neutron
This is the first part of the patch series which
 implements multiple ssh instance validation strategy.

This patch adds the new config options and have the effected
test classes to prepare the basic neutron resources for usage.

The default is using the fixed ips for connection, but
on the devstack side the configuration options are configured,
for neutron usage.

Change-Id: Ic5fc9bd1f7407d3430fcd33b03a226deed696d57
Implements: blueprint ssh-auth-strategy