Add unit tests for wait_for_resource_deletion
This commit adds unit tests to cover the wait_for_resource_deletion
method in the rest client. It covers the 3 cases for the method:
the resource is deleted before the timeout, the timeout is hit and an
exception is raised, and the service client child class doesn't have
a is_resource_deleted() method.

Partially implements bp unit-tests

Change-Id: I2acd1b5f44a7526c4b38d17d37bdc4b0f5db7d54