Validate get limits attributes of Nova V2 API
Now most attributes of Nova v2/v3 APIs are not checked in Tempest,
and this patch adds some tests which check these attributes to block
the backward incompatibility change in the future.

This patch adds the checks of "get limits" API responses.

The response body of v2 API is the following:
  {
    "limits": {
      "absolute": {
        "maxTotalRAMSize": 51200,
        "totalCoresUsed": 0,
        "maxTotalInstances": 10,
        "maxTotalFloatingIps": 10,
        "totalSecurityGroupsUsed": 0,
        "maxTotalCores": 20,
        "totalFloatingIpsUsed": 0,
        "maxSecurityGroups": 10,
        "maxServerMeta": 128,
        "maxPersonality": 5,
        "maxImageMeta": 128,
        "maxPersonalitySize": 10240,
        "maxSecurityGroupRules": 20,
        "maxTotalKeypairs": 100,
        "totalRAMUsed": 0,
        "totalInstancesUsed": 0
      },
      "rate": []
    }
  }

Partially implements blueprint nova-api-attribute-test

Change-Id: I503db2986ddcafb899f16d1547fd4857d355f2ed