Verify certificate API response attributes
Now most attributes of Nova v2/v3 APIs are not checked in Tempest,
and this patch adds some tests which check these attributes to block
the backward incompatibility change in the future.

This patch adds the checks of certificate API responses.
The response body of this API is the following:
{
    'certificate': {
        'data': 'Certificate:\n    Data:\n        Version:...',
        'private_key': '-----BEGIN RSA PRIVATE KEY-----\nM...'
    }
}

Partially implements blueprint nova-api-attribute-test

Change-Id: I5e97cfb5be56b4fa8de5d7262375b5b5b6de1b2a