fix cinder quota equality
The cinder quota test was of bad quality, and assumed that all
the functions in the class ran in linear order to work. If they
run in a different order the tenant could have additional quota
values beyond the strict defaults.

We can fix this by testing the returned quota contains the values
we're attempting to update.

Closes-Bug: #1291926

Change-Id: I53a154aac61368b7c20ac5703f3877fcf42f9781