Add config fixture support to unit tests
This commit adds a new env variable OS_TEST_LOCK_PATH to the
.testr.conf file to set the lock path for using oslo locks.
Previously, if an external lock couldn't be used to isolate test cases
because the unit tests do not have a config file. By getting a lock
path with an env variable and setting the location in the config
fixture external locks can be used by first invoking the config
fixture.

Partially implements bp unit-tests

Change-Id: I08c27ab125183f5d055eaa45c4f0e0a527445475