Fail a test if stack delete failed
Whenever a deleting stack goes to DELETE_FAILED this is an error which should fail
the test. Currently delete failures are ignored. As a side-effect of this, delete-failed
stacks are leaked even though tests pass.

The test may have done is own assertions on stack deletes, so a NotFound
exception is still passed.

Change-Id: I5f9e5901bcbd50bd8bd38c62175755ba59c6b837
Closes-Bug: #1295415