Add sahara to list of clients for T102
Change I6cb693298c75e400e5fe3a4bdd2532bca9d87b9b updated the list of
clients to include the missing official clients. However, at roughly
the same time savanna was renamed to sahara, so it wasn't included in
the list of clients. This commit corrects the oversight by adding
sahara to list of clients to check for T102.

Change-Id: Icfaad37b5f3b03f112ad6eb2fdedde38013287dd