orchestration API refactor test_list_resources
Rework test_list_resources so we do a more explicit test of the
resource list which is returned - we expect exactly one resource,
so using assertIn is less explicit than assertEqual.  Refactoring
the code which reads and validates the resource list will enable
other tests to more easily reuse the resource-listing code.

Change-Id: I0c1cb52d92f609781ee21dd3f7d52ceef85e0766