simplify heat test_limits
the test_limits test has a lot of conventions from other parts
of tempest which are not needed, like the use of setUpClass, and
setting _interface.

To do this we need to assign a client in the base class, further
refactoring should be able to remove a lot of code from these
tests.

This trims this down to the same functional equivalent, but with
less code.

Change-Id: Iabfa9afb7b55c4cdb86c519bce798e548463e7c4