Deduplicate negative test calls
Instead of duplicating the execute call for each negative test cases,
defining a generic way for all tests by registering a test function
dynamically on basis of the test class itself.

This is done using a class decorator (not inheritance) since it's
not possible to get the child class name on import time.

Change-Id: Ife0a8ea95ecfcac8a5f9c89babd74df0d5f884fe
Partially-implements: bp autogen-negative-tests