Split certificate API test
Test of certificate API was written as 1 test have 2 function tests.
There is no problem during it runs successfully.
But in case of failure, we cannot find easily about which part of function
was failed. It is not less preferred.
I split the test case as 1 test case checks 1 API.

Change-Id: I679394db42aad1e6e81bf3b023a69827aaf48d70