Inspect listed ports for created port with binding
Inspecting the binding attributes of unbound ports will cause a KeyError.
When tenant isolation is disabled, the subnet may contain unbound ports.
This updates the test to create a port that is bound to the host, and
asserts that the port exists in the listing and the listed port contains
the correct host_id in its binding attributes.  It no longer expects
the port_filter detail, which will be deprecated in Neutron eventually.

Change-Id: I43704407ca879ffd816644b65a03c615f978d432
Closes-bug: #1301782