Move exceptions back into one place
This commit removes the exceptions directory and reverts to the
previous organizational behavior of having a single exception file.
The directory was created with the intent of allowing additional
exceptions to be added as separate files. However this is just
distributing exceptions into separate files which would mean separate
imports whenever they were used. So new extensions were only being
added to the __init__ file to avoid the additional import, which is
essentially the previous behavior. So instead of adding the additional
directory it's easier just to leave everything in one file.

This is effectively a revert of 82411edb9c1089cac53572e7720559b927b59fbd

Change-Id: If15bd171b4708def30a650d514063e9795836746