Skip more pause tests if pause is not available
There are several compute API tests that attempt to pause a server but
should be skipped if the pause feature is not enabled.

Note that there are some negative tests which call the pause API but on
invalid server states, i.e. non-existent server, unpausing a server
that's not paused, etc, but those shouldn't get down to the virt driver
so we don't need to skip them since they should fail in the Nova API
service, not in the compute service.

Related-Bug: #1296858

Change-Id: I6f74c17af2c7540f913b71350770be7f004769a9