Consider state in firewall create/delete test
Firewall creation and deletion happens asynchronously.  It is not possible
to delete firewalls until they have become ACTIVE or to delete policies
or rules until its associated firewall has been completely deleted.

This updates the test_create_show_delete_firewall test to create the tenant
network resources required for an ACTIVE firewall and waits on the the
required state changes.  Also, this omits the 'status' key when comparing
the created firewall with the listed firewall as it is very much transient
during this phase of the test.

Change-Id: I556af59b5b22123229cb4d8ddcac9ce6a8b8f49c
Closes-bug: #1302942