Verify os-migration API response attributes
Now most attributes of Nova v2/v3 APIs are not checked in Tempest,
and this patch adds some tests which check these attributes to block
the backward incompatibility change in the future.

This patch adds the checks of v2/v3 os-migration API responses.

The response body of the API is the below:
 {
   "migrations": [
     {
       "id": 1,
       "status": "finished",
       "instance_uuid": "e9e4fdd7-f956-44ff-bfeb-d654a96ab3a2",
       "source_node": "localhost-dev",
       "source_compute": "localhost-dev",
       "dest_node": "localhost-dev",
       "dest_compute": "localhost-dev",
       "dest_host": "192.168.11.100",
       "old_instance_type_id": 6,
       "new_instance_type_id": 2,
       "created_at": "2014-03-18T00:14:49.000000",
       "updated_at": "2014-03-18T00:15:15.000000"
     }
   ]
 }

Partially implements blueprint nova-api-attribute-test

Change-Id: I21b7f55aa3310424f9c200c60a77e305af44cb48