Add load_test mechanism for InputScenarioUtils
Similar to what was done for the negative test framework this commit
adds a method to handle using CONF values with testscenarios.
Previously the attr's were set in the test class definition which
would force a getattr at import time. This will cause issues if the
config file is not present. By adding a method which won't be executed
at import but rather when the tests are run we can avoid this.

Change-Id: Ib635f8b895f6642f78a3af644170700f72259d46