Move the test_utils module to scenario test dir
The test_utils module in tempest.common.utils is only useful for
scenario tests. So it's a little confusing to have it in the common
directory. This commit corrects this by moving the file into the
scenario directory. The service tag requirement for scenario tests
hacking check is updated to only check scenario test files, because
a method in the new utils module was being picked up incorrectly.

Change-Id: I1d70745a5ee07923e680328e5b8a29bfb82d8099