Delete OrchestrationManager, and its unusual credentials
This removes the need for heat tests to be run with admin user
in the demo tenant.

NeutronResourcesTestJSON also needed to be modified as it was creating a router
with external_gateway_info which is enforced by policy as an admin-only operation.
This router is not required for connectivity in this test. An admin scenario
test can be created in the future which creates a router with
external_gateway_info.

Change-Id: Iba26845423b2bf52259a0c0cfab27ae09cfa4e60
Related-Bug: #1089261