fix test_compute_with_volumes
The test_integration_1 test was waiting on volume being in an
'in-use' state before proceeding. However the 'in-use' state
doesn't actually mean that the volume is attached, as 'attaching'
gets mapped to 'in-use' in nova.

This makes it possible for the test to proceed when the volume is
in an attaching state, and not actually ever attached. Which means
you can get to the detach phase on an non attached volume, and
thus explode.

Use the attachment status to filter out when then volume is not
actually ready.

In the process, make the test name something that's actually
descriptive for what it does.

Related-Bug: #1302774

Change-Id: Iaeb6a42a39b4adf8cb4bdff64efe00bf14821725