ssh_floating verify reboot
The reboot true option is usually useful for generate more ovs
events, but actually it can fail on the machine remains in SHUTOFF
state. A status check added, in order to help identify this situation.

The icmp echo test generated more logs than it required.
The intermediate pings will not be logged.

Change-Id: I524da1f2324227ef77755c070430221918e1694e