Remove {begin,roll}_detaching volume API tests
In preparation for changes to the cinder API in
https://review.openstack.org/#/c/90353/ , we need to make sure volumes
are in the expected state before we can call begin_detaching on it.

Adding this kind of check to the existing two tests would make them
scenario tests really, and also begin and roll detaching are already
tested as part of standard attach and detach scenario tests, plus are
kind of an implementation detail (even though, sadly, exposed as a
public API) and need not be tested by tempest tests.

Partial-bug: #1302774
Change-Id: I59e88344e83b8c0665a04c8ef55126a8f27bbf57