Make sure image exists in test
The heat bug/1312564 to add constraints to check whether image/keypair
exists for "AWS::AutoScaling::LaunchConfiguration" resource. So if the
image/keypair doesn't exist, the validate will fail. 

In tempest, there is a test which the "image" is invalid makes the 
Heat API test fail.

The property "ImageId" which defined in the non_empty_stack.yaml doesn't
exist. So to pass the parameter "image" to the property "ImageId" and
then bug/1312564 will successful in jenkins.

Related-Bug: #1312564
Change-Id: I4351930309c45c7fcb8a521ea0697b62153da5c7