Refactor _find_caller into a public test finder utility
This moves the _find_caller utility method from rest_client into it's
own public utility so that we can use it elsewhere in Tempest, like to
uniquely identify which test case fails in a TimeoutException.

This is useful to uniquely fingerprint timeout failures with
elastic-recheck which is not always possible right now because we don't
have multi-line query support with elastic-recheck.

Related-Bug: #1315201

Change-Id: I0cef7697e36a0ecd29c2540a5c283d322b2c62b7