Fix a few bugs when running with -u
The default conf file locator needed to reference the file another level up.

The config file set/write did not work if more than one option was changed.
I'm not sure why but think it would be better to just read once, set the
options, and write at the end anyway.

If we are updating and a service is there but not mentioned in the conf file,
we change the conf to True so should also add the service to the available
services.

Change-Id: I261f4ba08083cb7ed60c0a5bdd48a4dc16f4c23d