Make test_server_cfn_init a scenario test
Converting test_server_cfn_init to a scenario test is long overdue
since it is more of a scenario than an exercising of the heat API.

The heat-slow job has gone non-voting because this test fails
~%50 of the time due to bug #1297560. The server boot log is now
logged regardless of success or failure to build up data to diagnose
the issue.

This also adds several convenience functions to the test base class.

Co-Authored-By: Steven Hardy <shardy@redhat.com>
Related-Bug: #1297560
Change-Id: I077aeaf2bf8b292699eb20c5a75c59df35645913