Skip creation of network resources for disabled IPv6 tests
In some cases SetUpClass is called prior to determining that
the test case needs to be skipped, and that is okay. However,
for the IPv6 test cases, this may lead to errors as the very
creation of v6 network resources may not be supported by the
underlying CI infrastructure; therefore, for these cases let's
skip the test as the first step, if required.

Related-bug: #1282387

Change-Id: I84ae3f6d389c42393691718776bf1be18772aca6