tolerate HTTPNotFound in scenario manager teardown
When tearing down heat scenario tests, the tearDownClass fails if the
stack has already been deleted, because heatclient raises a HTTPNotFound
exception, so tolerate this in the teardown exception path.

This is required for certain tests where we explicitly delete the stack
from the test rather than relying on test cleanup, for example testing
deletion policy for volumes.

Change-Id: I1bc031ce00c845787311105eacb9afe4f779a3a5