Add "capabilities:" to multiple_backends extra-spec key
cinder-scheduler uses all the "capabilities:" to scope the
cinder-volume. With no "capabilities:" prefix set in extra_spec,
it will be no effect from cinder-scheduler to schedule the
cinder-volume, which is not the purpose of the testcase.
To address the review comment, This change will not remove but
adds 2 cases for the "capabilities:" prefix to volume type key:
extra-spec "capabilities:volume_backend_name".

Change-Id: Ie870091f2a5657b3ab6a0171de2cffe6945a0f7c
Closes-Bug: #1295050