Remove tempest_auto_config script
This script isn't being used by anyone to my knowledge nor has it
really been maintained since it was first added. We will be working
on a plan for automating tempest configuration soon and this
doesn't really fit the initial set requirements of requirements. It
also does not have any testing, so I expect that it has bit rotted
and probably doesn't work anymore.

Change-Id: Id896d9877312dfb6b12b54ea1c598a1940cdd9ce