Split large-ops into three tests so each has its own timeout
In the gate OS_TEST_TIMEOUT is 500 seconds, which is a per test timeout.
large-ops is run one test three times, so split it out to three tests so
each has its own OS_TEST_TIMEOUT.

This works because the cleanup of resources is done in tearDownClass and
tests inside a class are run serially (although order cannot be guaranteed).

Note: the image will be re-created for each test and destroyed on
tearDown from each test.

Change-Id: Iefa6f0fec9b3eea587b1437b47cce81e1a9c242c