Remove time.time mock from test_ssh
The function time.time is used in many places and especially in logging.
Mocking this function causes issues if the log level changes due to the
fact that every log message will cause a time.time call.

Splitting the test cases into 3 pieces since the time.sleep mock is
only usable for the good case. All tests with retries can't use it
and need to assert the result by measuring the time.

This fix is needed to use oslotest for tempest unit tests.

Change-Id: I800690f8ee1ee580cfcd71500b86782ff571ac0d
Closes-bug: bug 1320482