Add hacking import exception for service clients
This commit adds import exceptions to the service client imports.
The service clients are imported at the class level and it would be
a major effort to change this. However, we want to be enforcing H302
everywhere else. This commit adds an exception to allow importing the
service clients as classes.

Change-Id: Ife74539fe04e97740e45325140f407fd7476e1c7