Upload public image requires admin role by default
In glance the previous default was that any tenant could upload an
image as public which is likely not the desired default. This patch
changes the tempest tests so that images are no longer uploaded as
is_public=True by normal tenants so the policy change can merge in glance.

Glance change: https://review.openstack.org/#/c/92739/

Change-Id: If6a3aa0f7317b628b81b51d82108a18c701a1fd5