Don't store duplicate policies for server_group
It doesn't make sense to store same policies in a server_group.
We only need to store one and ignore the duplicate policies.

This patch relates to the bug I4f3ad544aef78cbbc076c7a47cca04832a2f5b4b
in Nova. So I need to skip one test-case here firstly in order to modify the
issue in Nova.

After the Nova's patch merged, this test-case will be restored,
and more correlate cases will definitly be supplied in tempest.

Change-Id: I26449a2a881be396daf75838451cfe01a915f513
Closes-Bug: #1324348