don't reuse servers in test_server_actions
One of the most problematic parts of Tempest failures has been
the test_server_actions test, which tries to optimize what it's
doing by reusing a functional server for subsequent tests.

The problem is, it's not clear that server is always functional.

It would be much better to take the time hit and just use a clean
server on every go if this increases reliability.

Change-Id: I78958250d85f2e6b32d912283936b2ee74c38feb