make logformat more similar to oslo
make the logformat actually match what we're used to reading in
the other logs, especially including the levels.

Because the logging fixture doesn't include the oslo logging
context processor, we are limited in what we can do here. A good
follow on would be to make a fixture which does understand oslo
logging processing so we could get request_id in the logs.

Change-Id: Ibb95aebd2df714fae331df13ad7789d6341af265