Use config generator rc instead of wrapper script
This commit removes the generate_sample wrapper script and adds an
rc file for the config generator to specify the alternate path.
This makes console messages from the generator tooling point
consistent with what is actually being used. Before because a wrapper
was used error messages would recommend running the incorrect tool.

Change-Id: I0ac792454777af6e3f1d718836031f9a14850229