Use the http image location conf option for api tests
This commit changes the api tests in test_images to use the http
image location from the config file instead of the hardcoded
example.com url. The example.com url would work in the case where
glance attempting to connect to the url using socket, does not raise
a socket error. However, this is dependent on the DNS configuration
of the host glance is running on because if example.com isn't
reachable glance will return a 400 because it can't get the image
size. This fixes the issue by using the location specified in config
which should be a valid location.

Closes-Bug: #1328738

Change-Id: I7cba95bd5f1462c33492869336ce9523a1545a07