Avoid snapshotting from compute authorization test
The compute authorization API tests currently rely on the creation of an image
via an instance snapshot.  Not all compute drivers support this and the tests
do not actually require the image to be a snapshot.  Instead, create a dummy
image in Glance directly to avoid making auth tests dependent on a hypervisor
feature.

Change-Id: I83a0fe12b1077ea87e402ec9897931631fa9cbf6
Closes-bug: #1331138