provide enough time to do expires in obj tests
When testing object expiration there are a lot of factors in
play, including the amount of time that rest calls take between
when we set the expiration, and when we eventually want to check.
If the time in the future is too close then our GET calls have
a real chance of 404ing because the object is gone.

So we should provide more slack to begin with (10 seconds) and
then a variable sleep time at the end to check objects 3 seconds
after they should have expired.

Closes-Bug: #1304119

Change-Id: I890bdbc14c926ba07f43d60ef5544ff37069d5df