Skip hypervisor uptime test for baremetal
Baremetal and ironic drivers have no concept of a hypervisor, thus
they do not implement the get_host_uptime() virt driver API call.
We would have to modify the underlying Tempest class that does the REST
calls to consider HTTP 501 (Not Implemented) successful, which would
affect the entire test suite. Instead, we just skip this test if
the baremetal/ironic driver is enabled.

Change-Id: I00c4146148063f6b20396dcbd0c911a386e488b1
Closes-Bug: #1338711