Base class for scenario test using Tempest clients
Define a very simple base class for migrating all scenario
tests to using tempest own clients.
This patch is the first step required to allow us to migrate
all tests one by one in dedicated changes, and finally drop
the unused Official Client pieces once migration is complete.

The various functions that are available in the OfficialClientTest
shall be migrated if / when needed by a specific test.

Change-Id: Ie82c8d8628ab50d6461a756713d56f78a2daa1dd
Partially-implements: bp:tempest-client-scenarios