Revert "Add tests for wait_for_server_status"
This unit test class added several race conditions to the unit tests as well as
added a dependency on having a valid tempest config file. At this point instead
of trying to fix all of these issues I'm pushing a revert in the short term and 
long-term we can just fix things and re-push these tests.

This reverts commit ae7c9ac4aac20a2832fe1beeaac0a688c6accbd9.

Change-Id: If6026611f3938be0716ce95491a8ee60a53e8c12
Closes-Bug: #1347096