Make javelin2 create_server wait for server to boot
Creating a server is a asynchronous command, make create_server wait for
the server to go to 'ACTIVE' state so we can run a check on the server
afterwards.

This also helps address the issue where the server isn't in ACTIVE yet
and grenade moves on and turns off nova-compute causing the server to go
to ERROR state.

Change-Id: Iafc0fe95e8224b1a800bd52aade3eb6488948639