Add network's status check after creation of network
Add assert for network status should be active

As per my understanding when we are creating network,
we can not say that we have created the same ..
till the point network is not ACTIVE and cannot be used.
And logically also , it should be like this only.
Keeping all this things in mind proposed the patch for the same.

Change-Id: Iceba14dde08fa38f942190131b82b9ba515ed0d6