Don't run lbaas scenario test as smoke test
Load Balancer as a service  is not part of neutron's primary
functionality, it is one of the advanced services that is not fully
baked yet. As part of the effort to only test advanced services against
neutron and not against all projects, drop this test from smoke.

Related patch: Ib315d30ff3c4463203f7ebe00e159a9680cf1616

Change-Id: I43239e9dc3e99e382ea12fb94f940f9029f4e488