Revert "Skip telemetry_notification_api test due to bug 1336755"
As https://review.openstack.org/#/c/114514/ has been merged, we can
try to check if changing of Ceilometer backend to the MongoDB makes the
situation better with running Ceilometer notification tests. Locally the
timeout issues were not reproduced both on MySQL and MongoDB, so this
change aims to check if the issue was solved by trying experimental runs
of MongoDB devstack Ceilometer testing.

This reverts commit 88004daa7b0fd746a60dd3b17d68ed4427b40e58.

Closes-Bug: #1276157

Change-Id: I41fce7e56a0e9fb111a95a266def786cb0f90c5e