Add a test for new-style container-sync
Add a file for testing new-style container synchronization. The file is named
as "test_container_sync_middleware.py" because new-style function is defined
in wsgi middleware. Add two config items for container-sync settings for
the new test.

Almost all test code is the same as "old-style" container sync test so new
test class overrides the old one. To reuse source code of the old class,
test_container_sync.py file is cleaned up.

!!!CONFIG DEFAULT VALUE IS CHANGED!!!
The default value of 'container_sync_timeout' item is changed from 120 to
600 because the default interval time when container-sync daemon process
runs is 300 sec so tests related to container-sync had a potential to
terminate in failure. From this patch, the default value is changed to
the double of the interval time for the purpose of safety.

Partially implements blueprint add-icehouse-swift-tests

Change-Id: I81c4ad6d452ed6a96ed81af3404f47cf802c41a0