Migrate cli test framework to tempest-lib
This commit switches the cli tests to use tempest-lib. The code which
was migrated to tempest-lib is removed and all the existing test cases
are converted to use the framework provided by the library. Future
improvements to the cli testing framework should be made in
tempest-lib after this commit.

Partially implements bp tempest-library
Change-Id: If2c9b3895a30905bc3d62b8daef783878d78399a