Make output from check_logs less verbose
This commit changes the output of check_logs.py to be a bit more
concise and clear. Instead of printing out every instance of errors
found in the log files it will only print out which individual files
had errors. If a file is marked as not allowing ERRORS or TRACES it
also prints that along with the filename. This should hopefully avoid
confusion around the operation of this script running in the gate.

Change-Id: Ib2a9a9736470cd3b445fb51a95a1a6c6696bb2cc