Properly detect grenade in check_logs script
The check logs script was not properly detecting grenade and would fail
grenade jobs if the logs were not clean. We don't want to do this
because the upgrade process is not log clean yet. Correct this by
setting grenade flag to true only if the DEVSTACK_GATE_GRENADE variable
has content.

Change-Id: Id8a339d78d981376b9af3a80aae8e4e18a68a85f