Fix TypeError in TestVolumeBootPattern
Commit eef192fc3c98a663f99f2fd944611bce0adffc0c changed
self._log_console_output() to self._log_console_output(self)
when dumping an instance's console output when ssh fails.

self in this case is the test class which is not an instance
(or list for that matter) so it blows up when we hit this
patch and masks the root failure.

Change-Id: Ieb68503a43269997a36c0a8b7f19f78ace97f251
Closes-Bug: #1363552