Only log console log if nova supports consoles
If a scenario test fails during ssh the standard procedure is to log
the console for the server. However, this can only be done if nova
supports consoles. This commit adds a check to see if nova supports
getting the console for a server before it attempts to do it.
Otherwise if nova doesn't support getting the console and an ssh
connection fails the test will fail on getting a console and not the
ssh connection. It looks like this check wasn't added during the
scenario test port to the tempest clients.

Change-Id: I681b3065d86cd31837a9ef5d91d25847fa47a8f4