Verify network connectivity before state check
This commit adds an initial ssh connection after bringing a server up
in setUp. This should ensure that the image has a chance to initialize
prior to messing with it's state. The test's here are to verify that
after performing a nova operation on a running instance network
connectivity is retained. However, it's is never checked that we can
connect to the server in the first place. A probable cause for the
constant ssh failures in these tests is that the server hasn't had a
finish it's cloud-init (or cirros-init) stage when we're stopping it,
this should also fix those issues.

Change-Id: I126fd4943582c4b759b3cc5a67babaa8d062fb4d
Partial-Bug: #1349617