Remove redundant waiter in create_volume cleanup
If wait_on_delete is True we are already waiting
for the volume delete, so the addCleanup_with_wait
call is redundant.

This just changes the addCleanup_with_wait call
to a normal addCleanup on the volume delete in
the case that wait_on_delete is True.

If wait_on_delete is False we still do the delete
but the wait is moved to the end of the test run
so it's asynchronous.

Change-Id: I6c4cd1a00ddc79de4c59c512b2d794e7380591f2
Partial-Bug: #1367857