Properly print STDOUT and STDERR in test_wrappers
If the unit tests in test_wrappers.py fail while using subprocess to
run the wrapper scripts STDOUT and STDERR are incorrectly used in the
exception method. Currently str() on the stringIO object is being
used for the message, which will only print information about the open
file. This commit fixes this by calling the read() method on the pipe
objects to print their contents.

Change-Id: I8fbf560fe2c706f1de96bb089730ec3f92fe3327