Removes bogus negative create image compute
Removes test_create_image_specify_multibyte_character_image_name
as it is not a valid test. It currently does not send an
invalid utf-8 sequence, but a valid unicode one. Also we can't actually
send an invalid utf-8 sequence via JSON as it is not possible to
encode an invalid UTF-8 sequence into a series of unicode characters
which is what JSON uses as strings. So its not possible for it
to fail this way. And we don't do XML testing for this type of test.

Change-Id: I39b5c017c8a173f8dd56a14c1bc7dc6952ac787d
Partially-Closes: 1006725