Make cli test namespace include test files
This commit adds the missing init files from the cli test suite. When
the test files were moved into subdirs as part of change
82e667ef4eaa79c20bbb7a528abda358edfb6a83 init files weren't included.
This meant that discovery couldn't find the tests and they weren't
runnable. Also, by enabling the tests unfixed paths also neede to be
updated in the heat tests, because they were referring to the
incorrect location.

Change-Id: I5807bb2f1f70cb0747c7831425e5df2bc48dd6ed