Lock test_resize_server_using_overlimit_* tests on 'compute_quotas'
The test_resize_server_using_overlimit_* negative tests are getting the
default quotas and creating a flavor that bumps one of the quota values
just over the limit, expecting an OverLimit failure when trying to
resize an instance to that temporary new flavor.

This can race with the os-quota-class-set admin test which changes the
default quotas, so lock the server negative tests using the same lock
fixture used in the quota class set tests.

Closes-Bug: #1372093

Change-Id: I380a7f057a6b6df94cc029f65f357de7368dc2f7