Orchestration use stack timeout waiting for ping response
The test_server_cfn_init scenario test boots a full Fedora image,
not the cirros image which is presumably used for compute tests
where the default 120s ping timeout may be appropriate.

To ensure we don't encounter any spurious failures due to booting the
much bigger orchestration image (which is Fedora in the gate), use
the stack timeout (default 900s) to override the ping timeout for
this test.

Change-Id: I4eba18333bdaa0ef41122cda85fd3f4d91b2cf5c
Closes-Bug: #1374175