Allow some lag for floating IP to reach a status
The current test checks the floating IP status as soon as possible
without allowing the status to converge.  This doesn't allow for the
fact that implementations don't report the new status until after
connectivity has been torn down.  There may be some expected lag that
is not allowed by the current test.

Change-Id: Icfe4bb2c40fde540bc2a6371630aa3f12dd945ff
Closes-Bug: #1381617