Use addCleanup when running swift scenario tests
Helper methods that create containers or objects are now responsible
for providing insurance that the container and object are removed.
delete_wrapper is used so that test cases may choose to delete objects
themselves and the resulting failure to cleanup in addCleanup will
not be an error.

Because deletion is a reasonable "basic op" scenario, the deletion
tests in test_swift_basic_ops are updated so they explicitly
validate deletion while also using addCleanup as a safety measure in
the event the test has unexpected failures.

Change-Id: I70b7172e9485cbd1eb881db3b2add3860da0935a