Remove "Disabled Reason" from test_cinder_service_list
The CLI test test_cinder_service_list checks the output of
"cinder service-list" for a set of column titles. One of the titles
it expects to find is "Disabled Reason". However, this title is not
guaranteed to be in the output; cinderclient only appends it if
the value shows up in the response from the server.

Change-Id: I060bad255ff17b132737a8878c26e7d5cf9dce82
Closes-Bug: #1392662