Adding api tests for dvr routers
in network_client.py:
update_router in json/network_client.py did not provide a way to update
the 'distributed' attribute of a router. Sending a rest call to a
system that does not support dvr (icehouse and older) or aa system with
dvr extension disabled will cause a failure, thus care must be taken to
only include the attribute in the rest call if it is explicitely
requested via the kwargs dict.

Tests Added:
-test_distributed_router_creation
-test_centralized_router_creation
-test_centralized_router_update_to_dvr

Change-Id: I19fb6fc260ac5aaf9b2a1456ade242bc3ee93b8d