Add internal methods for setting compute clients
If disabling XML tests on compute API tests by the config item
"xml_api_v2", we need to switch XML clients to JSON clients on
the other project API tests. But current code is difficult to
switch, because compute clients are many and the order is mess.
This patch adds internal methods to set compute clients for
cleanup.

Change-Id: I62bf48845c289b86f1af71e9fbf71e45b509a43d