Fix empty string as id in create flavor request
test_get_flavor_details_for_deleted_flavor specify "" as the
id value for creating the flavor.

Nova V2 API validation is loose and the API accepts this
value. However, V2.1 API will not allow empty string as id.

This patch fix this by passing 'None' as id in this test.

Change-Id: I86419368a4c85aaf039278999d3d8c3c1e224c74