Add check for too_slow_to_test flag on swift notification test
This commit adds a missing check for the too_slow_to_test flag on the
swift notification scenario test. On stable icehouse, without the
performance improvements from juno, this test is basically a crapshoot.
Checking if the environment is capable of running this test reliably
needs to be ensured prior to running the test.

Change-Id: I1c69341bfbd9b7540ca9cacae32dbaef2c526f93
Partial-Bug: #1394774