Skip test_volume_boot_pattern until bug 1373513 is fixed
Between the races to delete a volume and hitting timeouts because things
are hanging with lvm in Cinder and the various SSH timeouts, this test
is a constant burden.

The SSH problems have been around for a long time and don't seem to be
getting any new attention.

The Cinder volume delete hangs have also been around for awhile now and
don't seem to be getting much serious attention, so until the Cinder
volume delete hangs are fixed (or at least getting some serious
attention), let's just skip this test scenario.

Related-Bug: #1373513
Related-Bug: #1370496
Related-Bug: #1349617

Change-Id: Idb50bcdbc9683d322e9292abf50404e885a11a8e